requirements.txt
crewai
openai
streamlit
python-dotenv
pydantic